feat #340 : lineWidth 측정 로직 추가
- 최댓값, 최솟값을 적용할 최대거리 측정 필요
