+++ Outline/Outline/Source/Utils/Component/Card/SmallCarouselCard.swift
@@ -72,7 +72,7 @@ struct SmallCarouselCard: View {
                 }
                 .overlay(alignment: .bottom) {
                     VStack(spacing: 6) {
-                        CoordinatePathManager
+                        PathManager
                             .createPath(width: 200, height: 200, coordinates: data)
                             .scale(0.5)
                             .stroke(.customPrimary, style: StrokeStyle(lineWidth: 4, lineCap: .round, lineJoin: .round))
