+++ Outline/Outline/Source/Utils/Component/Card/SmallListCard.swift
@@ -70,7 +70,7 @@ struct SmallListCard: View {
                 }
                 .overlay(alignment: .bottom) {
                     VStack(spacing: 2) {
-                        CoordinatePathManager
+                        PathManager
                             .createPath(width: 100, height: 100, coordinates: data)
                             .scale(0.5)
                             .stroke(.customPrimary, style: StrokeStyle(lineWidth: 4, lineCap: .round, lineJoin: .round))
