+++ Outline/Outline/Source/View/NewRunning/CourseGuideView.swift
@@ -57,16 +57,16 @@ struct CourseGuideView: View {
 
 extension CourseGuideView {
     private var coursePath: some View {
-        CoordinatePathManager
+        PathManager
             .createPath(width: width, height: height, coordinates: coursePathCoordinates)
             .stroke(.customBlack.opacity(0.5), style: .init(lineWidth: 7, lineCap: .round, lineJoin: .round))
             .scaleEffect(0.8)
     }
     
     private var userPath: some View {
-        let canvasData = CoordinatePathManager.calculateCanvasData(coordinates: coursePathCoordinates, width: width, height: height)
+        let canvasData = PathManager.getCanvasData(coordinates: coursePathCoordinates, width: width, height: height)
         
-        return CoordinatePathManager
+        return PathManager
             .createPath(width: width, height: height, coordinates: userLocations, canvasData: canvasData)
             .stroke(.customPrimary, style: .init(lineWidth: 7, lineCap: .round, lineJoin: .round))
             .scaleEffect(0.8)
