+++ Outline/Outline/Source/View/PathTest/PathTestView.swift
@@ -13,19 +13,19 @@ struct PathTestView: View {
     @State private var coordinates: [Coordinate] = []
     
     private let parseManager = KMLParserManager()
-    private let width = 250.0
+    private let width = 200.0
     private let height = 200.0
     private let fileName = "fish"
     
     var body: some View {
         ZStack {
-                CoordinatePathManager.createPath(width: width, height: height, coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }))
+                PathManager.createPath(width: width, height: height, coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }))
                     .stroke(.green, style: .init(lineWidth: 4, lineCap: .round, lineJoin: .round) )
-                    .frame(width: CoordinatePathManager.calculateCanvasData(coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }), width: width, height: height).width, height: CoordinatePathManager.calculateCanvasData(coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }), width: width, height: height).height)
+                    .frame(width: PathManager.getCanvasData(coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }), width: width, height: height).width, height: PathManager.getCanvasData(coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }), width: width, height: height).height)
                     .overlay {
                         VStack {
-                            Text("\(CoordinatePathManager.calculateCanvasData(coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }), width: width, height: height).width)")
-                            Text("\(CoordinatePathManager.calculateCanvasData(coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }), width: width, height: height).height)")
+                            Text("\(PathManager.getCanvasData(coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }), width: width, height: height).width)")
+                            Text("\(PathManager.getCanvasData(coordinates: coordinates.map({ CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude) }), width: width, height: height).height)")
                         }
                     }
         }
